backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.36
alembic==1.14.0

# AI/ML Providers
google-generativeai==0.8.3
openai==1.55.3
anthropic==0.39.0

# TTS Providers
google-cloud-texttospeech==2.18.0
elevenlabs==1.11.0

# Content Ingestion
feedparser==6.0.11
beautifulsoup4==4.12.3
requests==2.32.3
yt-dlp==2024.12.23
youtube-transcript-api==1.2.4

# Video Processing
moviepy==1.0.3
pydub==0.25.1


# Utilities
httpx==0.28.1
python-multipart==0.0.20
websockets==14.1
psutil==7.2.1
python-json-logger==4.0.0
structlog==25.5.0
